+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [unreleased]
 
 * Add `--dry-run` option (for --tar-latest)
+* `--update` deletes backups from SQL that are not present in the borg repo
 
 ## [0.7]